Seperate my project into files with classes
